connect contract source identifier
