+++ protostar/cli/__init__.py
@@ -7,3 +7,4 @@ from .protostar_arg_type import create_map_protostar_type_name_to_parser
 from .protostar_argument import ProtostarArgument
 from .protostar_command import ProtostarCommand
 from .signable_command_util import SignableCommandUtil
+from .contract_source_identifier import create_contract_source_identifier_factory
