+++ protostar/cli/protostar_arg_type.py
@@ -1,10 +1,8 @@
-from pathlib import Path
 from typing import Any, Callable, Literal, Union
 
 from starkware.starknet.utils.api_utils import cast_to_felts
 
 from protostar.argument_parser import ArgTypeName, map_type_name_to_parser
-from protostar.compiler.project_compiler import ContractIdentifier
 from protostar.starknet_gateway import (
     SUPPORTED_BLOCK_EXPLORER_NAMES,
     Fee,
@@ -15,6 +13,8 @@ from protostar.starknet import Address
 from protostar.starknet.data_transformer import CairoOrPythonData
 from protostar.configuration_file import ConfigurationFile
 
+from .contract_source_identifier import create_contract_source_identifier_factory
+
 CustomProtostarArgTypeName = Literal[
     "felt",
     "wei",
@@ -23,7 +23,7 @@ CustomProtostarArgTypeName = Literal[
     "class_hash",
     "block_explorer",
     "input",
-    "contract_identifier",
+    "contract_source_identifier",
 ]
 
 ProtostarArgTypeName = Union[CustomProtostarArgTypeName, ArgTypeName]
@@ -47,8 +47,8 @@ def create_map_protostar_type_name_to_parser(configuration_file: ConfigurationFi
             return parse_block_explorer_type
         if argument_type == "input":
             return parse_input_arg_type
-        if argument_type == "contract_identifier":
-            return create_contract_identifier_parser(configuration_file)
+        if argument_type == "contract_source_identifier":
+            return create_contract_source_identifier_factory(configuration_file)
         return map_type_name_to_parser(argument_type)
 
     return map_protostar_type_name_to_parser
@@ -93,20 +93,3 @@ def parse_block_explorer_type(arg: str) -> SupportedBlockExplorerName:
     if arg not in SUPPORTED_BLOCK_EXPLORER_NAMES:
         raise ValueError()
     return arg
-
-
-def create_contract_identifier_parser(configuration_file: ConfigurationFile):
-    def parse_contract_identifier(arg: str) -> ContractIdentifier:
-        if arg.endswith(".cairo"):
-            contract_path = Path(arg)
-            if not contract_path.exists():
-                raise ValueError(
-                    f"The following contract doesn't exist: {contract_path}"
-                )
-            return contract_path
-        contract_names = configuration_file.get_contract_names()
-        if arg not in contract_names:
-            raise ValueError(f"Unknown contract: {arg}")
-        return arg
-
-    return parse_contract_identifier
