+++ protostar/composition_root.py
@@ -9,6 +9,9 @@ from protostar.cli import (
     MessengerFactory,
     ActivityIndicator,
 )
+from protostar.cli.contract_source_identifier import (
+    create_contract_source_identifier_factory,
+)
 from protostar.cli.lib_path_resolver import LibPathResolver
 from protostar.commands import (
     BuildCommand,
@@ -234,9 +237,7 @@ def build_di_container(
     argument_parser_facade = ArgumentParserFacade(
         protostar_cli,
         configuration_file,
-        parser_resolver=create_map_protostar_type_name_to_parser(
-            configuration_file=configuration_file
-        ),
+        parser_resolver=create_map_protostar_type_name_to_parser(configuration_file),
     )
 
     return DIContainer(protostar_cli, argument_parser_facade)
