+++ pyproject.toml
@@ -99,6 +99,7 @@ target-version = [
 ]
 
 [tool.isort]
+profile="black"
 lines_between_sections = 1
 only_sections = true
